requirements.txt
selenium
PyQt5
spotipy
configparser
pystray
pillow

winsdk; sys_platform == 'win32'
dbus-next; sys_platform == 'linux'